requirements.txt
alembic==1.15.2
annotated-types==0.7.0
asyncpg==0.30.0
dill==0.3.9
greenlet==3.1.1
inquirerpy==0.3.4
Mako==1.3.9
MarkupSafe==3.0.2
multiprocess==0.70.14
orjson==3.9.15
pfzy==0.3.4
prompt_toolkit==3.0.50
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
psycopg2; platform_system=="Windows"
psycopg2-binary; platform_system!="Windows"
robyn==0.65.0
rustimport==1.3.4
SQLAlchemy==2.0.40
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.1
watchdog==4.0.1
wcwidth==0.2.13
